refactor: move to galore_utils
